getIncludedDomainPurchase: do not return paid purchases
When a domain registration is bundled with a plan, its refund amount
should be 0. If the domain and plan have become unbundled, by, for
example, renewing the domain separately, then the domain should have its
own non-zero refund amount. In that case, we don't want to refund the
plan and the domain together.

This change checks the refund amount of the domain before considering it
bundled with a plan. If the amount is non-zero, it is considered separate.